MetadataStorage/app/requirements.txt
fastapi==0.111.0
pydantic_settings~=2.3
redis~=5.0